2009-12-26	Joel Sherrill <joel.sherrill@oarcorp.com>
	* t1lib-5.1.2/configure, t1lib-5.1.2/configure.in,
	t1lib-5.1.2/lib/t1lib/t1lib.h: When cross compiling you cannot
	execute autoconf probes. So hard code the 64-bit integer type for
	RTEMS.

+++ t1lib-5.1.2/lib/t1lib/t1lib.h
@@ -45,8 +45,8 @@ extern "C" {
 #define T1LIB_VERSIONSTRING      "5.1.2"
 
 /* Data Type information filled in by configure. */
-typedef short T1_int16;
-typedef int T1_int32;
+typedef int T1_int16;
+typedef long T1_int32;
 
 /* the data structure which contains the character bitmap description */
 typedef struct
